backend/requirements.txt
fastapi
uvicorn[standard]
reportlab
python-multipart
gspread
google-auth
